Library name conflict bug
PubSubClient is the name of the original knolleary library. 
The knolleary library is present on Arduino Library Manager, so the automatic library update system, that check only the name of a library, think that this is the knolleary library and ask to the user to update it (because the knolleary is at v2.0.3)
If the user accept the update this library is overwritten.

By this name change the Arduino Library Manager know that this is a different library and do not ask the update.